# Conflicts:
#	Cargo.toml
#	src/simulation/controller.rs
#	src/simulation/messaging/events.proto
#	src/simulation/messaging/events.rs
#	src/simulation/simulation.rs

+++ src/simulation/messaging/events.rs
@@ -191,7 +191,7 @@ impl Event {
         }
     }
 
-    pub fn new_travelled(person: u64, distance: f32, mode: u64) -> Event {
+    pub fn new_travelled(person: u64, distance: f64, mode: u64) -> Event {
         Event {
             r#type: Some(Travelled(TravelledEvent {
                 person,
